+++ Project_A/Assets/SCRIPTS/Respondent.cs
@@ -8,7 +8,7 @@ public class Respondent
     public float[] societalUtilities;
     
     // State
-    public int currentLS;      // 0-10 Scale
+    public float currentLS;      // 0-10 Scale (now a float, but is an int from ONS data)
     public int wealthTier;     // 0-10 Scale (For future policies like Wealth Tax)
 
     // Each person has an ID and two utility arrays, u_Self and u_Others
